fix(backend): appの初期化位置を変更
